+++ docs/src/pages/components/slider-styled/ContinuousSlider.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import styled from '@material-ui/core/styled';
+import styled from '@material-ui/core/experimentalStyled';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import Slider from '@material-ui/lab/SliderStyled';
