+++ packages/material-ui/src/experimentalStyled/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './experimentalStyled';
+export * from './experimentalStyled';
