+++ packages/material-ui/src/experimentalStyled/index.js
@@ -1,0 +1,1 @@
+export { default } from './experimentalStyled';
