+++ packages/material-ui/src/index.js
@@ -288,8 +288,8 @@ export * from './StepLabel';
 export { default as Stepper } from './Stepper';
 export * from './Stepper';
 
-export { default as styled } from './styled';
-export * from './styled';
+export { default as styled } from './experimentalStyled';
+export * from './experimentalStyled';
 
 export { default as SvgIcon } from './SvgIcon';
 export * from './SvgIcon';
