--- packages/material-ui/src/styled/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from './styled';
-export * from './styled';
