--- packages/material-ui/src/styled/index.js
@@ -1,1 +1,0 @@
-export { default } from './styled';
